Move ActionPlugin to MAVSDK project
